refactor: Simplify View method by adding ToTableListViewModel
Further simplified the View method by moving ViewModel construction logic to the Model.

Changes:
- Added `TableListViewModel` type to `internal/app/model.go`
- Added `ToTableListViewModel()` method to `app.Model`
- Updated `internal/views/table_list_view.go` to use `app.TableListViewModel`
- Simplified View method in `cmd/dito/main.go`:
  - Before: 14 lines for ScreenTableList case
  - After: 1 line using `m.ToTableListViewModel()`
- Removed unused `db` import from `internal/views/table_list_view.go`

Benefits:
- View method is now consistently simple across all screen cases
- ViewModel construction logic is encapsulated in the Model
- Better separation of concerns
- Improved maintainability
